+++ expression/builtin_json.go
@@ -17,10 +17,8 @@ import (
 	"github.com/juju/errors"
 	"github.com/pingcap/tidb/ast"
 	"github.com/pingcap/tidb/context"
-	"github.com/pingcap/tidb/mysql"
-	"github.com/pingcap/tidb/sessionctx/variable"
+	"github.com/pingcap/tidb/expression/xeval"
 	"github.com/pingcap/tidb/util/types"
-	"github.com/pingcap/tidb/util/types/json"
 	"github.com/pingcap/tipb/go-tipb"
 )
 
@@ -50,116 +48,6 @@ var (
 	_ functionClass = &jsonMergeFunctionClass{}
 )
 
-// argsAnyNull returns true if args contains any null.
-func argsAnyNull(args []types.Datum) bool {
-	for _, arg := range args {
-		if arg.Kind() == types.KindNull {
-			return true
-		}
-	}
-	return false
-}
-
-// datum2JSON gets or converts to JSON from datum.
-func datum2JSON(d types.Datum, sc *variable.StatementContext) (j json.JSON, err error) {
-	tp := types.NewFieldType(mysql.TypeJSON)
-	if d, err = d.ConvertTo(sc, tp); err == nil {
-		j = d.GetMysqlJSON()
-	}
-	return j, errors.Trace(err)
-}
-
-// parsePathExprs parses strings in datums into json.PathExpression.
-func parsePathExprs(datums []types.Datum) ([]json.PathExpression, error) {
-	pathExprs := make([]json.PathExpression, 0, len(datums))
-	for _, datum := range datums {
-		pathExpr, err := json.ParseJSONPathExpr(datum.GetString())
-		if err != nil {
-			return nil, errors.Trace(err)
-		}
-		pathExprs = append(pathExprs, pathExpr)
-	}
-	return pathExprs, nil
-}
-
-// createJSONFromDatums creates JSONs from Datums.
-func createJSONFromDatums(datums []types.Datum) ([]json.JSON, error) {
-	jsons := make([]json.JSON, 0, len(datums))
-	for _, datum := range datums {
-		// TODO: Here semantic of creating JSON from datum is the same with types.compareMysqlJSON.
-		// But it's different from "CAST semantic" because for string, cast will parse it into
-		// JSON but here and compareMysqlJSON needs a JSON with the string as primitives.
-		// We should rewrite this as a function for here and compareMysqlJSON both can use it.
-		var j json.JSON
-		switch datum.Kind() {
-		case types.KindNull:
-			j = json.CreateJSON(nil)
-		case types.KindMysqlJSON:
-			j = datum.GetMysqlJSON()
-		case types.KindInt64, types.KindUint64:
-			j = json.CreateJSON(datum.GetInt64())
-		case types.KindFloat32, types.KindFloat64:
-			j = json.CreateJSON(datum.GetFloat64())
-		case types.KindMysqlDecimal:
-			f64, err := datum.GetMysqlDecimal().ToFloat64()
-			if err != nil {
-				return jsons, errors.Trace(err)
-			}
-			j = json.CreateJSON(f64)
-		case types.KindString, types.KindBytes:
-			j = json.CreateJSON(datum.GetString())
-		default:
-			s, err := datum.ToString()
-			if err != nil {
-				return jsons, errors.Trace(err)
-			}
-			j = json.CreateJSON(s)
-		}
-		jsons = append(jsons, j)
-	}
-	return jsons, nil
-}
-
-// jsonModify is the portal for modify JSON with path expressions and values.
-func jsonModify(b baseBuiltinFunc, row []types.Datum, mt json.ModifyType) (j json.JSON, err error) {
-	args, err := b.evalArgs(row)
-	if err != nil {
-		return j, errors.Trace(err)
-	}
-	if argsAnyNull(args) {
-		return j, nil
-	}
-	j, err = datum2JSON(args[0], b.ctx.GetSessionVars().StmtCtx)
-	if err != nil {
-		return j, errors.Trace(err)
-	}
-
-	// alloc 1 extra element, for len(args) is an even number.
-	pes := make([]types.Datum, 0, (len(args)-1)/2+1)
-	vs := make([]types.Datum, 0, (len(args)-1)/2+1)
-	for i := 1; i < len(args); i++ {
-		if i&1 == 1 {
-			pes = append(pes, args[i])
-		} else {
-			vs = append(vs, args[i])
-		}
-	}
-	pathExprs, err := parsePathExprs(pes)
-	if err != nil {
-		return j, errors.Trace(err)
-	}
-	values, err := createJSONFromDatums(vs)
-	if err != nil {
-		return j, errors.Trace(err)
-	}
-
-	j, err = j.Modify(pathExprs, values, mt)
-	if err != nil {
-		return j, errors.Trace(err)
-	}
-	return j, nil
-}
-
 type jsonTypeFunctionClass struct {
 	baseFunctionClass
 }
@@ -177,15 +65,7 @@ func (b *builtinJSONTypeSig) eval(row []types.Datum) (d types.Datum, err error) 
 	if err != nil {
 		return d, errors.Trace(err)
 	}
-	if argsAnyNull(args) {
-		return d, nil
-	}
-	djson, err := datum2JSON(args[0], b.ctx.GetSessionVars().StmtCtx)
-	if err != nil {
-		return d, errors.Trace(err)
-	}
-	d.SetString(djson.Type())
-	return
+	return xeval.JSONType(args, b.ctx.GetSessionVars().StmtCtx)
 }
 
 type jsonExtractFunctionClass struct {
@@ -205,21 +85,7 @@ func (b *builtinJSONExtractSig) eval(row []types.Datum) (d types.Datum, err erro
 	if err != nil {
 		return d, errors.Trace(err)
 	}
-	if argsAnyNull(args) {
-		return d, nil
-	}
-	djson, err := datum2JSON(args[0], b.ctx.GetSessionVars().StmtCtx)
-	if err != nil {
-		return d, errors.Trace(err)
-	}
-	pathExprs, err := parsePathExprs(args[1:])
-	if err != nil {
-		return d, errors.Trace(err)
-	}
-	if djson1, found := djson.Extract(pathExprs); found {
-		d.SetMysqlJSON(djson1)
-	}
-	return
+	return xeval.JSONExtract(args, b.ctx.GetSessionVars().StmtCtx)
 }
 
 type jsonUnquoteFunctionClass struct {
@@ -239,19 +105,7 @@ func (b *builtinJSONUnquoteSig) eval(row []types.Datum) (d types.Datum, err erro
 	if err != nil {
 		return d, errors.Trace(err)
 	}
-	if argsAnyNull(args) {
-		return d, nil
-	}
-	djson, err := datum2JSON(args[0], b.ctx.GetSessionVars().StmtCtx)
-	if err != nil {
-		return d, errors.Trace(err)
-	}
-	unquoted, err := djson.Unquote()
-	if err != nil {
-		return d, errors.Trace(err)
-	}
-	d.SetString(unquoted)
-	return
+	return xeval.JSONUnquote(args, b.ctx.GetSessionVars().StmtCtx)
 }
 
 type jsonSetFunctionClass struct {
@@ -267,12 +121,11 @@ func (c *jsonSetFunctionClass) getFunction(args []Expression, ctx context.Contex
 }
 
 func (b *builtinJSONSetSig) eval(row []types.Datum) (d types.Datum, err error) {
-	j, err := jsonModify(b.baseBuiltinFunc, row, json.ModifySet)
+	args, err := b.evalArgs(row)
 	if err != nil {
 		return d, errors.Trace(err)
 	}
-	d.SetMysqlJSON(j)
-	return
+	return xeval.JSONSet(args, b.ctx.GetSessionVars().StmtCtx)
 }
 
 type jsonInsertFunctionClass struct {
@@ -288,12 +141,11 @@ func (c *jsonInsertFunctionClass) getFunction(args []Expression, ctx context.Con
 }
 
 func (b *builtinJSONInsertSig) eval(row []types.Datum) (d types.Datum, err error) {
-	j, err := jsonModify(b.baseBuiltinFunc, row, json.ModifyInsert)
+	args, err := b.evalArgs(row)
 	if err != nil {
 		return d, errors.Trace(err)
 	}
-	d.SetMysqlJSON(j)
-	return
+	return xeval.JSONInsert(args, b.ctx.GetSessionVars().StmtCtx)
 }
 
 type jsonReplaceFunctionClass struct {
@@ -309,12 +161,11 @@ func (c *jsonReplaceFunctionClass) getFunction(args []Expression, ctx context.Co
 }
 
 func (b *builtinJSONReplaceSig) eval(row []types.Datum) (d types.Datum, err error) {
-	j, err := jsonModify(b.baseBuiltinFunc, row, json.ModifyReplace)
+	args, err := b.evalArgs(row)
 	if err != nil {
 		return d, errors.Trace(err)
 	}
-	d.SetMysqlJSON(j)
-	return
+	return xeval.JSONReplace(args, b.ctx.GetSessionVars().StmtCtx)
 }
 
 type jsonMergeFunctionClass struct {
@@ -334,17 +185,5 @@ func (b *builtinJSONMergeSig) eval(row []types.Datum) (d types.Datum, err error)
 	if err != nil {
 		return d, errors.Trace(err)
 	}
-	if argsAnyNull(args) {
-		return d, nil
-	}
-	jsons := make([]json.JSON, 0, len(args))
-	for _, arg := range args {
-		j, err := datum2JSON(arg, b.ctx.GetSessionVars().StmtCtx)
-		if err != nil {
-			return d, errors.Trace(err)
-		}
-		jsons = append(jsons, j)
-	}
-	d.SetMysqlJSON(jsons[0].Merge(jsons[1:]))
-	return
+	return xeval.JSONMerge(args, b.ctx.GetSessionVars().StmtCtx)
 }
